Conflicts:
	modules/juce_audio_basics/buffers/juce_AudioSampleBuffer.cpp
	modules/juce_gui_basics/native/juce_mac_NSViewComponentPeer.mm

+++ modules/juce_gui_basics/native/juce_ios_Windowing.mm
@@ -312,12 +312,12 @@ bool Desktop::canUseSemiTransparentWindows() noexcept
     return true;
 }
 
-Point<int> MouseInputSource::getCurrentRawMousePosition()
+Point<float> MouseInputSource::getCurrentRawMousePosition()
 {
     return juce_lastMousePos;
 }
 
-void MouseInputSource::setRawMousePosition (Point<int>)
+void MouseInputSource::setRawMousePosition (Point<float>)
 {
 }
