+++ modules/juce_gui_extra/native/juce_mac_SystemTrayIcon.cpp
@@ -99,21 +99,21 @@ public:
 
             if (isLeft || isRight)  // Only mouse up is sent by the OS, so simulate a down/up
             {
-                owner.mouseDown (MouseEvent (mouseSource, Point<int>(),
+                owner.mouseDown (MouseEvent (mouseSource, Point<float>(),
                                              eventMods.withFlags (isLeft ? ModifierKeys::leftButtonModifier
                                                                          : ModifierKeys::rightButtonModifier),
                                              &owner, &owner, now,
-                                             Point<int>(), now, 1, false));
+                                             Point<float>(), now, 1, false));
 
-                owner.mouseUp (MouseEvent (mouseSource, Point<int>(), eventMods.withoutMouseButtons(),
+                owner.mouseUp (MouseEvent (mouseSource, Point<float>(), eventMods.withoutMouseButtons(),
                                            &owner, &owner, now,
-                                           Point<int>(), now, 1, false));
+                                           Point<float>(), now, 1, false));
             }
             else if (type == NSMouseMoved)
             {
-                owner.mouseMove (MouseEvent (mouseSource, Point<int>(), eventMods,
+                owner.mouseMove (MouseEvent (mouseSource, Point<float>(), eventMods,
                                              &owner, &owner, now,
-                                             Point<int>(), now, 1, false));
+                                             Point<float>(), now, 1, false));
             }
         }
     }
